Moved extensions to rollbar_common
