+++ rollbar_common/lib/rollbar_common.dart
@@ -1,3 +1,12 @@
 library rollbar_common;
 
+export 'src/extension/function.dart';
+export 'src/extension/object.dart';
+export 'src/extension/collection.dart';
+export 'src/extension/math.dart';
+export 'src/extension/date_time.dart';
+export 'src/extension/environment.dart';
+
+export 'src/extension/tuple.dart';
+
 export 'src/connectivity_monitor.dart';
