+++ rollbar_common/lib/src/extension/function.dart
