+++ rollbar_common/lib/src/extension/math.dart
