+++ rollbar_common/lib/src/extension/object.dart
