+++ rollbar_common/lib/src/extension/tuple.dart
