--- rollbar_dart/lib/rollbar_dart.dart
@@ -1,1 +1,0 @@
-export 'rollbar.dart';
