+++ rollbar_dart/lib/src/data/payload/frame.dart
@@ -1,8 +1,8 @@
 import 'package:meta/meta.dart';
 import 'package:stack_trace/stack_trace.dart' as trace;
+import 'package:rollbar_common/rollbar_common.dart';
 
 import '../../ext/trace.dart';
-import '../../ext/collection.dart';
 
 /// Contains the information of a single frame in a stack trace.
 @sealed
