+++ rollbar_dart/lib/src/payload_record_database.dart
@@ -2,10 +2,9 @@ import 'dart:collection';
 
 import 'package:meta/meta.dart';
 import 'package:sqlite3/sqlite3.dart';
+import 'package:rollbar_common/rollbar_common.dart';
 
-import 'ext/object.dart';
 import 'ext/database.dart';
-import 'ext/collection.dart';
 import 'ext/identifiable.dart';
 
 import 'data/payload_record.dart';
