+++ rollbar_flutter/lib/src/connectivity_monitor.dart
@@ -1,8 +1,7 @@
 import 'dart:async';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
-
-import 'package:rollbar_dart/rollbar_dart.dart';
+import 'package:rollbar_dart/rollbar.dart';
 
 /// Service aiding in optimizing network operations.
 /// [FlutterConnectivityMonitor] is designed to work as a singleton.
